This is a business school program ranking webiste.
To show our data, we build a desktop website to let users search ranking
information.
The frontend of our website is run on Bootstrap and jQuery. Bootstrap is
modular and consists essentially of a series of Less stylesheets that
implement the various components of the toolkit.  jQuery is designed to
make it easier to navigate a document, select DOM elements, create
animations, handle events.
The backend is Node.js, Express.js. Node.js allows the creation of Web
servers and networking tools using JavaScript and a collection of
"modules" that handle various core functionality. Express.js, or simply
Express, is a web application framework for Node.js.